Cloud VM: mount data to /var/lib/docker (#8)
Since Nextcloud AIO manages itself and uses docker volumes, the easiest way to persisty all data in a separate disk is to just mount it to `/var/lib/docker`. This is still an easier long-term solution than managing the entire stack manually - we can always migrate (relatively) easily in the future anyway.

Reviewed-on: https://git.koval.net/cyclane/kovalhome/pulls/8